# Add slice support for *_argnum parameters