@jakevdp 2025-05-12
Closing due to lack of CLA, and also because this looks LLM-generated.